requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cache2==0.2.2
flower==2.0.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pendulum==3.0.0
pillow==11.0.0
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
pyasn1==0.6.1
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pytest==8.3.4
pytest-asyncio==0.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==3.2.1
python-multipart==0.0.19
pytz==2024.2
redis==4.6.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
time-machine==2.16.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
